% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harpPoint_tidyverse.R
\name{arrange.harp_fcst}
\alias{arrange.harp_fcst}
\title{Arrange columns from tables in a \code{harp_fcst} object.}
\usage{
\method{arrange}{harp_fcst}(.fcst, ...)
}
\arguments{
\item{.fcst}{A harp_fcst object}

\item{...}{Arguments as in \link[dplyr]{arrange}}
}
\description{
Works in the same way as \link[dplyr]{arrange}, except runs on all tables in
the \code{harp_fcst} object. This means that only common columns between the
objects can safely be arranged.
}

