% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_hexbin.R
\name{bin_fcst_obs}
\alias{bin_fcst_obs}
\alias{bin_fcst_obs.harp_det_point_df}
\alias{bin_fcst_obs.harp_ens_point_df}
\title{Create a bi-variate histogram of forecast, observation pairs}
\usage{
bin_fcst_obs(
  .fcst,
  parameter,
  groupings = "lead_time",
  num_bins = 30,
  show_progress = TRUE,
  ...
)

\method{bin_fcst_obs}{harp_det_point_df}(
  .fcst,
  parameter,
  groupings = "lead_time",
  num_bins = 30,
  show_progress = TRUE,
  fcst_model = NULL,
  ...
)

\method{bin_fcst_obs}{harp_ens_point_df}(
  .fcst,
  parameter,
  groupings = "lead_time",
  num_bins = 30,
  show_progress = TRUE,
  fcst_model = NULL,
  ...
)
}
\arguments{
\item{.fcst}{A `harp_df` data frame or a `harp_list`.}

\item{parameter}{The column containing the parameter. Can be unquoted, a
quoted string, or an embraced variable name (i.e {{var}}).}

\item{groupings}{The groupings for which to compute the binned densities.
Must be a vector of strings, or a list of vectors of strings.}

\item{num_bins}{The number of bins into which to partition the observations.}

\item{show_progress}{Logical. Whether to show a progress bar.}

\item{...}{Arguments for methods.}

\item{fcst_model}{The name of the forecast model. If `.fcst` does not
contain a `fcst_model`, a new column is created and populated with this
value. If a `fcst_model` column exists, the value in the column is replaced
with this value.}
}
\description{
Values of forecasts and observations are binned into bands whereby the
density of forecast, observation pairs for each bin is calculated. Under the
hood, the data are binned into hexagons using \code{\link[hexbin]{hexbin}}.
Hexagons are used since they have symmetry with nearest neighbours unlike
square bins, and at plot time they are the polygon with the maximum number of
sides that tessellate.
}
