% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_point_verif.R
\name{bind_point_verif}
\alias{bind_point_verif}
\title{Bind harp verification objects into a single object}
\usage{
bind_point_verif(...)
}
\arguments{
\item{...}{Verification objects that are to be combined}
}
\description{
Bind harp verification objects into a single object
}
