% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_score.R
\name{bootstrap_score}
\alias{bootstrap_score}
\title{Bootstrap a score}
\usage{
bootstrap_score(.fcst, score_function, parameter, n,
  groupings = "leadtime", confidence_interval = 0.95, ...)
}
\arguments{
\item{.fcst}{A \code{harp_fcst} object.}

\item{parameter}{The parameter that gives the name of the observations
columns.}

\item{n}{Th number of bootstrap replicants.}

\item{groupings}{The groups to compute verification scores for. The default
is leadtime.}

\item{...}{Other arguments to the score function. e.g. thresholds.}

\item{score}{The name of the verification function to bootstrap for.}
}
\description{
Compute the confidence intervals of verification scores using a bootstrap
method. If there is more than one forecast the confidence for the differences
between forecasts is also computed.
}
