% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harpPoint_tidyverse.R
\name{dplyr_list}
\alias{dplyr_list}
\alias{mutate_list}
\alias{filter_list}
\alias{select_list}
\title{dplyr verbs for lists}
\usage{
mutate_list(.list, ...)

filter_list(.list, ...)

select_list(.list, ...)
}
\arguments{
\item{.list}{A list of data frames}

\item{...}{Other arguments to the dplyr verb}
}
\description{
When you have a list of data frames, such as the output to a verification
function, you may want to wrangle data in those data frames at the same time.
This can be achieved using the dplyr verb followed by _list. For data frames
where the function is applicaple the modified data frame is returned. If the
verb fails (e.g. because the specified columns don't exist), the data frame
is silently returned unmodified
}
\seealso{
/link[dplyr]{mutate}, /link[dplyr]{filter}, /link[dplyr]{select}
}
