% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_score_funcs.R
\name{ens_brier}
\alias{ens_brier}
\title{Brier Score for an ensemble forecast}
\usage{
ens_brier(
  .fcst,
  parameter,
  thresholds,
  comparator = c("ge", "gt", "le", "lt", "eq", "between", "outside"),
  include_low = TRUE,
  include_high = TRUE,
  groupings = "lead_time",
  climatology = "sample",
  num_ref_members = NA,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{.fcst}{A \code{harp_df} or \code{harp_list} object with tables that have a column
for observations, or a single forecast table.}

\item{parameter}{The name of the column for the observations data. Can be the
column name, quoted, or unquoted. If a variable it should be embraced -
i.e. wrapped in \code{{{}}}.}

\item{thresholds}{A numeric vector of thresholds for which to compute the
threshold based scores. Set to NULL (the default) to only compute summary
scores.}

\item{groupings}{The groups for which to compute the scores. See
\link[dplyr]{group_by} for more information of how grouping works. Should
be a character vector, or a list of character vectors. If a list each
combination of groups in the list is computed separately for each score.
See \link[harpCore]{make_verif_groups} for a helper function for making
more complex grouping structures.}

\item{climatology}{The climatology to use for the Brier Skill Score. Can be
"sample" for the sample climatology (the default), a named list with
elements eps_model and member to use a member of an eps model in the
harp_fcst object for the climatology, or a data frame with columns for
threshold and climatology and also optionally lead_time.}

\item{num_ref_members}{For "fair" scores, the score is scaled to be valid for
this number of ensemble members. Set to NA (the default) to not modify the
score.}

\item{show_progress}{Logical - whether to show progress bars. Defaults to
\code{TRUE}.}

\item{...}{Reserved for methods.}
}
\value{
A \code{harp_verif} object with an entry for "ens_threshold_scores".
}
\description{
Computes the Brier Score, its decomposition, and the Brier Skill Score for an
ensemble forecast. If \code{num_ref_members} is supplied, the fair Brier Score is
also computed for that number of members. For the Brier Skill Score, the
default is to use the sample climatology as the reference. It is also
possible to pass the climatology as a data frame that can be inner joined
with the forecast data.
}
\details{
This is simply a wrapper around \code{\link{ens_verify()}} with \code{brier}
as the only selected score.
}
