% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_probabilities.R
\name{ens_probabilities}
\alias{ens_probabilities}
\title{Compute probabilities of threshold exceedence for ensemble forecasts}
\usage{
ens_probabilities(
  .fcst,
  thresholds,
  comparator = c("ge", "gt", "le", "lt", "eq", "between", "outside"),
  include_low = TRUE,
  include_high = TRUE,
  parameter = NULL
)
}
\arguments{
\item{.fcst}{A \code{harp_df} or \code{harp_list} object with tables that have a column
for observations, or a single forecast table.}

\item{thresholds}{A numeric vector of thresholds for which to compute the
threshold based scores. Set to NULL (the default) to only compute summary
scores.}

\item{parameter}{The name of the column for the observations data. Can be the
column name, quoted, or unquoted. If a variable it should be embraced -
i.e. wrapped in \code{{{}}}.}
}
\value{
A \code{harp_list} object with each data frame having columns for threshold,
fcst_prob and obs_prob instead of the columns for each member forecast.
}
\description{
Compute probabilities of threshold exceedence for ensemble forecasts
}
