% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_score_funcs.R
\name{ens_rank_histogram}
\alias{ens_rank_histogram}
\title{Compute the rank histogram an ensemble forecast}
\usage{
ens_rank_histogram(
  .fcst,
  parameter,
  groupings = "lead_time",
  jitter_fcst = NULL,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{.fcst}{A \code{harp_df} or \code{harp_list} object with tables that have a column
for observations, or a single forecast table.}

\item{parameter}{The name of the column for the observations data. Can be the
column name, quoted, or unquoted. If a variable it should be embraced -
i.e. wrapped in \code{{{}}}.}

\item{groupings}{The groups for which to compute the scores. See
\link[dplyr]{group_by} for more information of how grouping works. Should
be a character vector, or a list of character vectors. If a list each
combination of groups in the list is computed separately for each score.
See \link[harpCore]{make_verif_groups} for a helper function for making
more complex grouping structures.}

\item{jitter_fcst}{A function to perturb the forecast values by. This is used
to account for observation error in the rank histogram. For other
statistics it is likely to make little difference since it is expected that
the observations will have a mean error of zero.}

\item{show_progress}{Logical - whether to show progress bars. Defaults to
\code{TRUE}.}

\item{...}{Reserved for methods.}
}
\value{
A \code{harp_verif} object with an entry for "ens_summary_scores"
}
\description{
The rank histogram as a nested column in a \code{harp_verif} object. By default
the score is aggregated by lead time, though any other grouping strategy can
be chosen. Observation errors can be taken into account by supplying a
jitter function. See \code{\link{jitter_fcst}} for more details.
}
\details{
This is simply a wrapper around \code{\link{ens_verify()}} with \code{rank_hist}
as the only selected score.
}
