% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_read_and_verify.R
\name{ens_read_and_verify}
\alias{ens_read_and_verify}
\title{Title}
\usage{
ens_read_and_verify(start_date, end_date, parameter, fcst_model, fcst_path,
  obs_path, lead_time = seq(0, 48, 3), num_iterations = length(lead_time),
  thresholds = NULL, members = NULL, obsfile_template = "obstable",
  groupings = "leadtime", by = "1d", climatology = "sample",
  stations = NULL, jitter_fcst = NULL, gross_error_check = TRUE,
  min_allowed = NULL, max_allowed = NULL, show_progress = FALSE,
  verif_path = NULL)
}
\arguments{
\item{start_date}{Start date to for the verification. Should be numeric or
character. YYYYMMDD(HH)(mm).}

\item{end_date}{End date for the verification. Should be numeric or
character.}

\item{parameter}{The parameter to verify.}

\item{fcst_model}{The forecast model(s) to verify. Can be a single string or
a character vector of model names.}

\item{fcst_path}{The path to the forecast FCTABLE files.}

\item{obs_path}{The path to the observatoin OBSTABLE files.}

\item{lead_time}{The lead times to verify.}

\item{num_iterations}{The number of iterations per verification calculation.
The default is to do the same number of iterations as there are lead times.
If a small number of iterations is set, it may be useful to set
\code{show_progress = TRUE}. The higher the number of iterations, the
smaller the amount of data that is held in memory at any one time.}

\item{thresholds}{The thresholds to compute categorical scores for.}

\item{members}{The members to include in the iteration. This will select the
same member numbers from each \code{fcst_model}. In the future it will
become possible to specify members for each \code{fcst_model}.}

\item{obsfile_template}{The template for OBSTABLE files - the default is
"obstable", which is \code{OBSTABLE_{YYYY}.sqlite}.}

\item{groupings}{The groups to verify for. The default is "leadtime". Another
common grouping might be \code{groupings = c("leadtime", "fcst_cycle")}.}

\item{by}{The frequency of forecast cycles to verify.}

\item{climatology}{The climatology to use for the Brier Skill Score. Can be
"sample" for the sample climatology (the default), a named list with
elements eps_model and member to use a member of an eps model in the
harp_fcst object for the climatology, or a data frame with columns for
threshold and climatology and also optionally leadtime.}

\item{stations}{The stations to verify for. The default is to use all
stations from \link[harpIO]{station_list} that are common to all
\code{fcst_model} domains.}

\item{jitter_fcst}{A function to perturb the forecast values by. This is used
to account for observation error in the rank histogram. For other
statistics it is likely to make little difference since it is expected that
the observations will have a mean error of zero.}

\item{gross_error_check}{Logical of whether to perform a gross error check.}

\item{min_allowed}{The minimum value of observation to allow in the gross error
check. If set to NULL the default value for the parameter is used.}

\item{max_allowed}{The maximum value of observation to allow in the gross error
check. If set to NULL the default value for the parameter is used.}

\item{show_progress}{Logical - whether to show a progress bar. Defaults to
FALSE.}

\item{verif_path}{If set, verification files will be saved to this path.}
}
\value{
A list containting two data frames: \code{ens_summary_scores} and
  \code{ens_threshold_scores}.
}
\description{
Title
}

