% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_score_funcs.R
\name{ens_spread_and_skill}
\alias{ens_spread_and_skill}
\title{Compute the skill (RMSE) and spread of an ensemble forecast}
\usage{
ens_spread_and_skill(
  .fcst,
  parameter,
  groupings = "lead_time",
  circle = NULL,
  spread_drop_member = NULL,
  jitter_fcst = NULL,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{.fcst}{A \code{harp_df} or \code{harp_list} object with tables that have a column
for observations, or a single forecast table.}

\item{parameter}{The name of the column for the observations data. Can be the
column name, quoted, or unquoted. If a variable it should be embraced -
i.e. wrapped in \code{{{}}}.}

\item{groupings}{The groups for which to compute the scores. See
\link[dplyr]{group_by} for more information of how grouping works. Should
be a character vector, or a list of character vectors. If a list each
combination of groups in the list is computed separately for each score.
See \link[harpCore]{make_verif_groups} for a helper function for making
more complex grouping structures.}

\item{circle}{If set the parameter is assumed to be cyclic for bias
calculations. Should be this distance around a circle in the units of the
parameter, so would typically have a value of 360 for degrees or \code{2 * pi}
for radians.}

\item{spread_drop_member}{Which members to drop for the calculation of the
ensemble variance and standard deviation. For harp_fcst objects, this can
be a numeric scalar - in which case it is recycled for all forecast models;
a list or numeric vector of the same length as the harp_fcst object, or a
named list with the names corresponding to names in the harp_fcst object.}

\item{jitter_fcst}{A function to perturb the forecast values by. This is used
to account for observation error in the rank histogram. For other
statistics it is likely to make little difference since it is expected that
the observations will have a mean error of zero.}

\item{show_progress}{Logical - whether to show progress bars. Defaults to
\code{TRUE}.}

\item{...}{Reserved for methods.}
}
\value{
A \code{harp_verif} object with an entry for "ens_summary_scores"
}
\description{
The ensemble mean and spread are computed as columns in a \code{harp_verif}
object. By default the scores are aggregated over lead time by other grouping
variables cam be chosen. The mean bias and standard deviation of the error
are also computed.
}
\details{
This is simply a wrapper around \code{\link{ens_verify()}} with \code{summary} as
the only selected score.
}
