% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_score_funcs.R
\name{ens_tw_crps}
\alias{ens_tw_crps}
\title{Threshold weighted CRPS for an ensemble forecast}
\usage{
ens_tw_crps(
  .fcst,
  parameter,
  thresholds,
  comparator = c("ge", "gt", "le", "lt", "between"),
  include_low = TRUE,
  include_high = TRUE,
  groupings = "lead_time",
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{.fcst}{A \code{harp_df} or \code{harp_list} object with tables that have a column
for observations, or a single forecast table.}

\item{parameter}{The name of the column for the observations data. Can be the
column name, quoted, or unquoted. If a variable it should be embraced -
i.e. wrapped in \code{{{}}}.}

\item{thresholds}{A numeric vector of thresholds for which to compute the
threshold based scores. Set to NULL (the default) to only compute summary
scores.}

\item{groupings}{The groups for which to compute the scores. See
\link[dplyr]{group_by} for more information of how grouping works. Should
be a character vector, or a list of character vectors. If a list each
combination of groups in the list is computed separately for each score.
See \link[harpCore]{make_verif_groups} for a helper function for making
more complex grouping structures.}

\item{show_progress}{Logical - whether to show progress bars. Defaults to
\code{TRUE}.}

\item{...}{Reserved for methods.}
}
\value{
A \code{harp_verif} object with an entry for "ens_threshold_scores".
}
\description{
Computes the threshold weighted CRPS for an ensemble forecast. First values
outside of the threshold are clamped to the threhshold value and then the
CRPS is computed using the kernel method.
}
\details{
This is simply a wrapper around \code{\link{ens_verify()}} with \code{tw_crps}
as the only selected score.
}
