% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_date.R
\name{expand_date}
\alias{expand_date}
\title{Expand date columns into year, month, day, hour, minute.}
\usage{
expand_date(.fcst, date_col)
}
\arguments{
\item{.fcst}{A harp_fcst object or a data frame with a column with the string
'date' in its name.}

\item{date_col}{The unquoted name of the date column that you wish to expand.}
}
\value{
An object of the same type as '.fcst' with extra columns for year,
  month, day, hour, minute. The extra colun names are prefixed with
  'date_col' with the string 'date' removed.
}
\description{
Expand date columns into year, month, day, hour, minute.
}
