% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_data.R
\name{fake_point_data}
\alias{fake_point_data}
\title{Make fake point data}
\usage{
fake_point_data(
  fcst_dttm,
  fcst_model,
  lead_time = seq(0, 48),
  SID = seq(1000, 1100),
  members = NULL,
  seed = NULL
)
}
\arguments{
\item{fcst_dttm}{A vector of date-times, for example as produced by
\code{\link[harpCore]{seq_dttm}}.}

\item{fcst_model}{A vector of forecast model names.}

\item{lead_time}{A vector of lead times.}

\item{SID}{A vector of station IDs.}

\item{members}{Either \code{NULL} (the default) to generate deterministic data,
or a numeric vector of member numbers to generate ensemble data.}

\item{seed}{A seed to generate random numbers. By setting this to a numeric
value, the generated data are fully reproducible.}
}
\description{
This function generates fake point data to be used for testing functions, or
for making reproducible examples. For each \code{fcst_dttm} and \code{SID} a time
series of forecasts is produced using a random walk. Observations are
generated by adding small random numbers to the forecast values. For ensemble
data, each member is generated by adding small random perturbations to the
first generated forecast.
}
\examples{
# Generate deterministic data for 2 stations for 1 model
fake_point_data(20240901, "a", lead_time = seq(0, 3), SID = c(1, 2))

# Do the same but for 3 models
fake_point_data(20240901, letters[1:3], lead_time = seq(0, 3), SID = c(1, 2))

# Generate ensemble data for 4 members + control
fake_point_data(
  20240901, "a", lead_time = seq(0, 3), SID = c(1, 2), members = seq(0, 4)
)

}
