% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fcprob}
\alias{fcprob}
\title{Compute the forecast probabilities for given thresholds}
\usage{
fcprob(fc, thresholds)
}
\arguments{
\item{fc}{A two dimensional array of EPS data with members in columns.}

\item{byrow}{The thresholds to compute probabilities for}
}
\description{
Compute the forecast probabilities for given thresholds
}

