% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harpPoint_tidyverse.R
\name{filter.harp_fcst}
\alias{filter.harp_fcst}
\title{Filter a \code{harp_fcst} object.}
\usage{
\method{filter}{harp_fcst}(.fcst, ...)
}
\arguments{
\item{.fcst}{a harp_fcst object.}

\item{...}{Arguments as in \link[dplyr]{filter}}
}
\description{
Works on each table in the \code{harp_fcst} object in the same way as \link[dplyr]{filter}
}
