% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_members.R
\name{gather_members}
\alias{gather_members}
\title{Convert EPS forecast data from wide format data frame to long format data
frame.}
\usage{
gather_members(
  .fcst,
  member_regex = "_mbr[[:digit:]]+$|_mbr[[:digit:]]+_lag[[:graph:]]*$"
)
}
\arguments{
\item{.fcst}{An EPS forecast data frame in wide format.}

\item{member_regex}{Regular expression for column names that contain forecasts for a
single member.}
}
\value{
An EPS data frame in long format.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{\link[harpCore]{pivot_members}} is now the preferred method for
transforming between long and wide data frames since it supports classes
that were introduced in version 0.1.0.
}
