% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_crps.R
\name{harp_crps}
\alias{harp_crps}
\title{Compute the CRPS for ensemble prediction systems.}
\usage{
harp_crps(.fcst)
}
\arguments{
\item{.fcst}{A FCST object with column names: mname, member, SID, fcdate,
leadtime, validdate, forecast, obs.}
}
\value{
A list with named elements: CRPS, CRPSpot, Reli, alpha, beat,
  heaviside0, heavisideN.
}
\description{
Compute the CRPS for ensemble prediction systems.
}

