% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harpPoint-package.R
\docType{package}
\name{harpPoint-package}
\alias{harpPoint}
\alias{harpPoint-package}
\title{harpPoint: Point verifition for NWP forecasts}
\description{
Functions for computing verification scores for NWP forecasts. Part of the harp ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/harphub/harpPoint}
  \item Report bugs at \url{https://github.com/harphub/harpPoint/issues}
}

}
\author{
\strong{Maintainer}: Andrew Singleton \email{andrewts@met.no}

Authors:
\itemize{
  \item Alex Deckmyn \email{alex.deckmyn@meteo.be}
}

}
\keyword{internal}
