% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harpPoint_tidyverse.R
\name{join_models}
\alias{join_models}
\title{Join all models into a single ensemble.}
\usage{
join_models(.fcst, join_type = "inner", name = "joined_models", ...)
}
\arguments{
\item{.fcst}{A harp_fcst object with any multimodel data merged with
\link{merge_multimodel}.}

\item{join_type}{The type of join to perform. See \link[dplyr]{join}.}

\item{name}{The name of the resulting model.}

\item{...}{Other arguments to \link[dplyr]{join}.}
}
\value{

}
\description{
The function is most useful for finding common cases between models.
}
