% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_to_fcst.R
\name{join_to_fcst}
\alias{join_to_fcst}
\title{Combine data with point forecasts.}
\usage{
join_to_fcst(.fcst, .join, join_type = c("inner", "left", "right",
  "full", "semi", "anti"), ...)
}
\arguments{
\item{.fcst}{A \code{harp_fcst} object.}

\item{.join}{A data frame to join to each table in the \code{harp_fcst}
object.}

\item{join_type}{How to join the data frame. Acceptable values are: "inner",
"left", "right", "full", "semi", "anti". See \code{\link[dplyr]{join}} for
more details.}

\item{by}{Which columns to join by.}
}
\value{
The input forecast data fram with an "obs" column added.
}
\description{
Works in a similar way to the \link[dplyr]{join} family of functions, excpet
that the join type is specified as an argument.
}
\examples{

}
