% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_forecast.R
\name{lag_forecast}
\alias{lag_forecast}
\title{Lags a forecast}
\usage{
lag_forecast(.fcst, fcst_model, parent_cycles)
}
\arguments{
\item{.fcst}{A harp_fcst object.}

\item{fcst_model}{The name of the forecast model in the harp_fcst object to
be lagged. Must be quoted.}

\item{parent_cycles}{A numeric vector of forecast cycles that form the
parents for the lagging. Members between parent cycles will be added to the
parent cycle that is the closest that comes after the child cycle.}
}
\value{
A harp_fcst object with \code{fcst_model} now containing the lagged
  forecast.
}
\description{
Lagging is done by supplying the parent forecast cycles. This function will
then work out which cycles belong to the parent and return the lagged
forecast with members from the child cycles added to the parent cycle.
}
