% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_validdate.R
\name{last_validdate}
\alias{last_validdate}
\title{Return the last valid date in a forecast}
\usage{
last_validdate(.fcst)
}
\arguments{
\item{.fcst}{A \code{harp_fcst} object or a table containing a column named
validdate with the data in unix time format.}
}
\value{
The first valid time in YYYYMMDDhhmm format for the input object.
}
\description{
The function is intended to be used to find the last date for which to fetch
observations against which to verify.
}
