% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_pools.R
\name{make_bootstrap_pools}
\alias{make_bootstrap_pools}
\title{Make pools for block bootstrapping}
\usage{
make_bootstrap_pools(.fcst, pool_col, pool_length, overlap = FALSE)
}
\arguments{
\item{.fcst}{A \code{harp_fcst} object}

\item{pool_col}{The column used to define the pools. Can be the column name,
quoted, or unquoted. If a variable it should be embraced - i.e. wrapped in
\code{{{}}}}

\item{pool_length}{The length of a pool. Numeric or a character with a unit
qualifier if \code{pool_col} is in date-time format. The unit qualifier can
be : "s" = seconds, "m" = minutes, "h" = hours, "d" = days.}

\item{overlap}{Logical. Whether the pools should overlap.}
}
\value{
A data frame with columns from \code{pool_col} and "pool".
}
\description{
For doing a block bootstrap using \link{bootstrap_verify}, the blocks can be
passed as a data frame with a "pool" column telling \link{bootstrap_verify}
how to pool the data into blocks. \code{make_bootstrap_pools} is a function
to make such a data frame.
}
\details{
Typically block bootstrapping would be used if there are serial
auto-correlations in the data. If for example auto-correlations are suspected
between forecasts, pools could be defined from the \code{fcdate} column to
create blocks of data where those auto-correlations are maintained.

Pools may be set to overlap, whereby a new pool is created beginning at each
new value in \code{pool_col}. The length of a pool should be defined in the
units used in \code{pool_col} - if \code{pool_col} is a date-time column,
then \code{pool_length} is assumed to be in hours, though the units can be
set by adding a qualifier letter: "s" = seconds, "m" = minutes, "h" = hours,
"d" = days.
}
\examples{
make_bootstrap_pools(ens_point_df, lead_time, 2)
make_bootstrap_pools(ens_point_df, lead_time, 2, overlap = TRUE)

# pool_col as a variable
my_col <- "lead_time"
make_bootstrap_pools(ens_point_df, {{my_col}}, 2)
}
