% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_multimodel.R
\name{merge_multimodel}
\alias{merge_multimodel}
\title{Merge sub models of a multimodel ensemble into a single ensemble}
\usage{
merge_multimodel(.fcst, keep_sub_models = TRUE)
}
\arguments{
\item{.fcst}{An object of class \code{harp_fcst} as read in by
\link{read_point_forecast}}

\item{keep_sub_models}{Set to FALSE to discard the sub models as separate
elements in the \code{harp_fcst} list. The default behaviour is to keep
them.}
}
\value{
A \code{harp_fcst} object with one layer - each element is a table of
forecast data for a model.
}
\description{
When multimodel ensembles are read in only the sub models are stored. If the
whole ensemble is needed, this function will merge all of the sub models into
a single ensemble. Note that there is no renaming of the members so the sub
model names are retained. If there are no multi model ensembles, the input is
silently returned unaltered.
}
