% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_obs.R
\name{merge_obs}
\alias{merge_obs}
\title{Combine observations with point forecasts.}
\usage{
merge_obs(FCST, OBS, join_type = c("inner", "left", "right", "full", "semi",
  "anti"), by = c("SID", "validdate"))
}
\arguments{
\item{FCST}{A point forecast data frame.}

\item{OBS}{A point observations data frame.}

\item{join_type}{How to join the data frame. Acceptable values are: "inner",
"left", "right", "full", "semi", "anti". See \code{\link[dplyr]{join}} for
more details.}

\item{by}{Which columns to join by. Default is c("SID", "validdate").}
}
\value{
The input forecast data fram with an "obs" column added.
}
\description{
Combine observations with point forecasts.
}

