% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harpPoint_tidyverse.R
\name{mutate.harp_fcst}
\alias{mutate.harp_fcst}
\title{Mutate columns from tables in a \code{harp_fcst} object.}
\usage{
\method{mutate}{harp_fcst}(.fcst, ...)
}
\arguments{
\item{.fcst}{A harp_fcst object}

\item{...}{Arguments as in \link[dplyr]{mutate}}
}
\description{
Works in the same way as \link[dplyr]{mutate}, except runs on all tables in
the \code{harp_fcst} object.
}
