% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harpPoint_tidyverse.R
\name{mutate_at.harp_fcst}
\alias{mutate_at.harp_fcst}
\title{Mutate selected columns from tables in a \code{harp_fcst} object.}
\usage{
mutate_at.harp_fcst(.fcst, .mutate_vars, .mutate_funs, ...)
}
\arguments{
\item{.fcst}{A harp_fcst object}

\item{...}{Arguments as in \link[dplyr]{mutate}}
}
\description{
Works in the same way as \link[dplyr]{mutate_at}, except runs on all tables
in the \code{harp_fcst} object.
}
