% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfect_score.R
\name{perfect_score}
\alias{perfect_score}
\title{Set or modify and output the perfect scores for different verificaiton
metrics}
\usage{
perfect_score(...)
}
\arguments{
\item{...}{Named arguments that give values for the perfect score}
}
\value{
A named vector
}
\description{
The function returns a named vector giving the perfect score for all
verification scores computed by harpPoint that give a single value. These
perfect values can be modified by specifying them in named arguments and
perfect scores for new scores can be set in the same way.
}
\examples{
perfect_score()
perfect_score(bss = 1)
perfect_score(bias = -1)
}
