% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_verification.R
\name{plot_verification}
\alias{plot_verification}
\title{Title}
\usage{
plot_verification(plotData, Score, parameter, xAxis = "leadtime",
  yAxis = NULL, colourBy = "mname", facetBy = NULL, linetypeBy = NULL,
  plotPoints = TRUE, ncol = 2, colourTable = NULL, plotCaption = NULL,
  includeZero = TRUE, xLabel = NULL, yLabel = NULL, linesize = 1.1,
  pointsize = 2, num_legend_rows = 3, legend.position = "bottom",
  plot_title = NULL, plot_subtitle = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}

