% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled_bootstrap_score.R
\name{pooled_bootstrap_score}
\alias{pooled_bootstrap_score}
\title{Bootstrapping by pools}
\usage{
pooled_bootstrap_score(
  .fcst,
  score_function,
  parameter,
  n,
  pooled_by = "fcdate",
  groupings = "leadtime",
  confidence_interval = 0.95,
  min_cases = 25,
  perfect_scores = perfect_score(),
  show_progress = TRUE,
  ...
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
