% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harpPoint_tidyverse.R
\name{pull.harp_fcst}
\alias{pull.harp_fcst}
\title{Pull columns from tables in a \code{harp_fcst} object.}
\usage{
\method{pull}{harp_fcst}(.fcst, ...)
}
\arguments{
\item{.fcst}{A harp_fcst object}

\item{...}{Arguments as in \link[dplyr]{pull}}
}
\description{
Works in the same way as \link[dplyr]{pull}, except runs on all tables in
the \code{harp_fcst} object. This means that only common columns between the
objects can safely be pulled.
}
