% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_stations.R
\name{pull_stations}
\alias{pull_stations}
\title{Pull station IDs from forecast or observations}
\usage{
pull_stations(.fcst)
}
\arguments{
\item{.fcst}{A harp_fcst object or data frame}
}
\value{
A vector of station IDs
}
\description{
Pull station IDs from forecast or observations
}
