% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rankHistogram}
\alias{rankHistogram}
\title{Compute the rank histogram for an EPS}
\usage{
rankHistogram(obs, df)
}
\arguments{
\item{obs}{A vector of observations.}

\item{fc}{A two dimensional array of EPS data with members in columns.}
}
\description{
Compute the rank histogram for an EPS
}
