% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harpPoint_tidyverse.R
\name{rename.harp_fcst}
\alias{rename.harp_fcst}
\title{Rename columns from tables in a \code{harp_fcst} object.}
\usage{
\method{rename}{harp_fcst}(.fcst, ...)
}
\arguments{
\item{.fcst}{A harp_fcst object}

\item{...}{Arguments as in \link[dplyr]{rename}}
}
\description{
Works in the same way as \link[dplyr]{rename}, except runs on all tables in
the \code{harp_fcst} object. This means that only common columns between the
objects can safely be renamed.
}
