% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harpPoint_tidyverse.R
\name{rename_with.harp_fcst}
\alias{rename_with.harp_fcst}
\title{Rename multiple columns from tables in a \code{harp_fcst} object.}
\usage{
\method{rename_with}{harp_fcst}(.fcst, ...)
}
\arguments{
\item{.fcst}{A harp_fcst object}

\item{...}{Arguments as in \link[dplyr]{rename_with}}
}
\description{
Works in the same way as \link[dplyr]{rename_with}, except runs on all tables
in the \code{harp_fcst} object. This means that only common columns between
the objects can safely be renamed.
}
