% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roc}
\alias{roc}
\title{Compute ROC and area under ROC}
\usage{
roc(obs, pred, thresholds)
}
\arguments{
\item{obs}{A vector observations (value 0,1)}

\item{pred}{A vector of probabilities \link{0,1}.}

\item{thresholds}{A vector of threshold probabilities}
}
\value{
a list with area and vectors thresholds, H, F
}
\description{
Compute ROC and area under ROC
}
