% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harpPoint_tidyverse.R
\name{select.harp_fcst}
\alias{select.harp_fcst}
\title{Select columns from tables in a \code{harp_fcst} object.}
\usage{
\method{select}{harp_fcst}(.fcst, ...)
}
\arguments{
\item{.fcst}{A harp_fcst object}

\item{...}{Arguments as in \link[dplyr]{select}}
}
\description{
Works in the same way as \link[dplyr]{select}, except runs on all tables in
the \code{harp_fcst} object. This means that only common columns between the
objects can safely be selected.
}

