% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_members.R
\name{select_members}
\alias{select_members}
\title{Select members from a harp_fcst object}
\usage{
select_members(.fcst, members, include_lagged_members = TRUE)
}
\arguments{
\item{.fcst}{A harp_fcst object}

\item{members}{The members to select. Can be a numeric vector, or a named
list to select members from spcific forecast models in the harp_fcst
object.}
}
\value{
A harp_fcst object.
}
\description{
Select members from a harp_fcst object
}
