% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sort2d}
\alias{sort2d}
\title{Sort a 2d array.}
\usage{
sort2d(x, byrow = TRUE)
}
\arguments{
\item{x}{A two dimensional numeric array.}

\item{byrow}{Set to true sort rows, false to sort columns.}
}
\description{
Sort a 2d array.
}

