% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sort_members_df}
\alias{sort_members_df}
\title{Sort a 2d array from a data frame.}
\usage{
sort_members_df(df, byrow = TRUE)
}
\arguments{
\item{byrow}{Set to true sort rows, false to sort columns.}

\item{x}{A two dimensional numeric array.}
}
\description{
Sort a 2d array from a data frame.
}
