% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_members.R
\name{spread_members}
\alias{spread_members}
\title{Convert EPS forecast data from long format data frame to wide format data
frame.}
\usage{
spread_members(.fcst, ...)
}
\arguments{
\item{.fcst}{An EPS forecast data frame in long format.}
}
\value{
An EPS data frame in wide format.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{\link[harpCore]{pivot_members}} is now the preferred method for
transforming between long and wide data frames since it supports classes
that were introduced in version 0.1.0.
}
