% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harpPoint_tidyverse.R
\name{transmute.harp_fcst}
\alias{transmute.harp_fcst}
\title{Transmute columns from tables in a \code{harp_fcst} object.}
\usage{
\method{transmute}{harp_fcst}(.fcst, ...)
}
\arguments{
\item{.fcst}{A harp_fcst object}

\item{...}{Arguments as in \link[dplyr]{transmute}}
}
\description{
Works in the same way as \link[dplyr]{transmute}, except runs on all tables in
the \code{harp_fcst} object. This means that only common columns between the
objects can safely be arranged.
}
