% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_det.R
\name{verify_det}
\alias{verify_det}
\title{Verify a deterministic forecast}
\usage{
verify_det(FCST, thresholds = NULL)
}
\arguments{
\item{FCST}{A harp deterministc forecast object with an obs column.}

\item{thresholds}{A numeric vector of thresholds for which to compute
threshold based scores. If set to NULL, no threshold based scores will be
computed.}
}
\value{
A list with two data frames: \cr \code{scores} - for non-threshold
  based scores, \cr\code{scores_thresh} - for threshold based scores.
}
\description{
Verify a deterministic forecast
}

