% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_bootstrap_score.R
\name{bind_bootstrap_score}
\alias{bind_bootstrap_score}
\title{Bind output of bootstrap_score / pooled_bootstrap_score}
\usage{
bind_bootstrap_score(...)
}
\arguments{
\item{...}{A list of outputs from `bootstrap_score()` /
`pooled_bootstrap_score()`}
}
\value{
An object of class `harp_bootstrap`
}
\description{
`r lifecycle::badge("deprecated")`
}
