% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_point_verif.R
\name{bind_point_verif}
\alias{bind_point_verif}
\title{Bind harp verification objects into a single object}
\usage{
bind_point_verif(...)
}
\arguments{
\item{...}{Verification objects that are to be combined. Can be individual
objects or a list.}
}
\value{
A harp verification object with all attributes moved up to columns
and combined in the new attributes.
}
\description{
For plotting it may be desirable to combine various different verifications
into a single object. For example, verification for different parameters for
the same set of forecast models.
}
