% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_cases.R
\name{common_cases}
\alias{common_cases}
\title{Title}
\usage{
common_cases(.fcst)
}
\arguments{
\item{.fcst}{A harp3 format data frame of forecasts}
}
\value{
The input data frame with only the common stations and forecast dates
  for each experiment selected.
}
\description{
Title
}
