% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/det_probabilities.R
\name{det_probabilities}
\alias{det_probabilities}
\title{Compute binary probabilities for deterministic forecasts}
\usage{
det_probabilities(.fcst, parameter, thresholds, obs_probabilities = TRUE)
}
\arguments{
\item{.fcst}{A \code{harp_fcst} object with tables that have a column for
observations, or a single forecast table.}

\item{parameter}{The name of the column for the observed data.}

\item{thresholds}{A numeric vector of thresholds for which to compute
probabilities.}

\item{obs_probabilities}{A logical indicating whether or not to compute the
binary probabilities for the observations.}
}
\value{
A \code{harp_fcst} object with each data frame having columns for threshold,
fcst_prob and optionally obs_prob instead of the raw forecast column.
}
\description{
Compute binary probabilities for deterministic forecasts
}
