% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/det_verify.R
\name{det_verify}
\alias{det_verify}
\title{Compute verification scores for deterministic forecasts.}
\usage{
det_verify(.fcst, parameter, thresholds = NULL, groupings = "leadtime",
  show_progress = FALSE)
}
\arguments{
\item{.fcst}{A \code{harp_fcst} object with tables that have a column for
observations, or a single forecast table.}

\item{parameter}{The name of the column for the observed data.}

\item{thresholds}{A numeric vector of thresholds for which to compute the
threshold based scores. Set to NULL (the default) to only compute summary
scores.}

\item{groupings}{The groups for which to compute the scores. See
\link[dplyr]{group_by} for more information of how grouping works.}

\item{show_progress}{Logical - whether to show a progress bar. The default is
FALSE.}
}
\value{
A list containting two data frames: \code{det_summary_scores} and
  \code{det_threshold_scores}.
}
\description{
Compute verification scores for deterministic forecasts.
}
