% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_brier.R
\name{ens_brier}
\alias{ens_brier}
\title{Brier score and its decomposition for an ensemble.}
\usage{
ens_brier(
  .fcst,
  parameter,
  thresholds,
  groupings = "leadtime",
  climatology = "sample",
  num_ref_members = NA,
  keep_score = c("both", "brier", "reliability"),
  show_progress = FALSE
)
}
\arguments{
\item{.fcst}{A \code{harp_fcst} object with tables that have a column for
observations, or a single forecast table.}

\item{parameter}{The name of the column for the observed data.}

\item{thresholds}{A numeric vector of thresholds for which to compute the
Brier Score.}

\item{groupings}{The groups for which to compute the ensemble mean and
spread. See \link[dplyr]{group_by} for more information of how grouping
works.}

\item{climatology}{The climatology to use for the Brier Skill Score. Can be
"sample" for the sample climatology (the default), a named list with
elements eps_model and member to use a member of an eps model in the
harp_fcst object for the climatology, or a data frame with columns for
threshold and climatology and also optionally leadtime.}

\item{num_ref_members}{The number of members for which to compute the fair
Brier score.}

\item{keep_score}{\code{ens_brier} computes the Brier Score and its
components, as well as reliability. If you only want to output the Brier
score and its decomposition set to "brier", to only keep the reliability
information set to "reliability". To keep both, set to "both". The default
behaviour is "both".}

\item{...}{Used internally depending on the class of the input.}
}
\value{
A data frame with data grouped for the \code{groupings} column(s) and
  columns for \code{brier_score}, \code{brier_skill_score} and the
  deomposition of the brier score - \code{brier_score_reliability},
  \code{brier_score_resolution} and \code{brier_score_uncertainty}.
}
\description{
Brier score and its decomposition for an ensemble.
}
