% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_brier.R
\name{ens_brier}
\alias{ens_brier}
\title{Title}
\usage{
ens_brier(.fcst, parameter, thresholds, groupings = "leadtime")
}
\arguments{
\item{.fcst}{A \code{harp_fcst} object with tables that have a column for
observations, or a single forecast table.}

\item{parameter}{The name of the column for the observed data.}

\item{thresholds}{A numeric vector of thresholds for which to compute the
Brier Score.}

\item{groupings}{The groups for which to compute the ensemble mean and
spread. See \link[dplyr]{group_by} for more information of how grouping
works.}
}
\value{
A data frame with data grouped for the \code{groupings} column(s) and
  columns for \code{brier_score}, \code{brier_skill_score} and the
  deomposition of the brier score - \code{brier_score_reliability},
  \code{brier_score_resolution} and \code{brier_score_uncertainty}.
}
\description{
Title
}

