% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_crps.R
\name{ens_crps}
\alias{ens_crps}
\title{Continuous Rank Probability Score (CRPS) for an ensemble.}
\usage{
ens_crps(.fcst, parameter, groupings = "leadtime", keep_full_output = FALSE)
}
\arguments{
\item{.fcst}{A \code{harp_fcst} object with tables that have a column for
observations., or a single forecast table.}

\item{keep_full_ouput}{}
}
\description{
The CRPS and its decomposition are computed as columns in a \code{harp_fcst}
object. Typically the scores are aggregated over lead time, but other grouping
variables cam be chosen.
}

