% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_score_funcs.R
\name{ens_crps_decomp}
\alias{ens_crps_decomp}
\title{Compute the CRPS and its decomposition of an ensemble forecast}
\usage{
ens_crps_decomp(
  .fcst,
  parameter,
  groupings = "lead_time",
  jitter_fcst = NULL,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{.fcst}{A \code{harp_df} or \code{harp_list} object with tables that have a column
for observations, or a single forecast table.}

\item{parameter}{The name of the column for the observations data. Can be the
column name, quoted, or unquoted. If a variable it should be embraced -
i.e. wrapped in \code{{{}}}.}

\item{groupings}{The groups for which to compute the scores. See
\link[dplyr]{group_by} for more information of how grouping works.}

\item{jitter_fcst}{A function to perturb the forecast values by. This is used
to account for observation error in the rank histogram. For other
statistics it is likely to make little difference since it is expected that
the observations will have a mean error of zero.}

\item{show_progress}{Logical - whether to show progress bars. Defaults to
\code{TRUE}.}

\item{...}{Reserved for methods.}
}
\value{
A \code{harp_verif} object with an entry for "ens_summary_scores"
}
\description{
The CRPS is computed using the CDF method along with its decomposition into
reliability and potential components as columns in a \code{harp_verif}
object. By default the score is aggregated by lead time, though any other
grouping strategy can be chosen. Note that this method of computing the
CRPS is typically slower and more memory intensive than the kernel method
used by \code{\link{ens_crps()}}.
}
\details{
This is simply a wrapper around \code{\link{ens_verify()}} with \code{crps_decomp}
as the only selected score.
}
