% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_score_funcs.R
\name{ens_hexbin}
\alias{ens_hexbin}
\title{Compute hexbins for an ensemble forecast}
\usage{
ens_hexbin(
  .fcst,
  parameter,
  groupings = "lead_time",
  verify_members = TRUE,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{.fcst}{A \code{harp_df} or \code{harp_list} object with tables that have a column
for observations, or a single forecast table.}

\item{parameter}{The name of the column for the observations data. Can be the
column name, quoted, or unquoted. If a variable it should be embraced -
i.e. wrapped in \code{{{}}}.}

\item{groupings}{The groups for which to compute the scores. See
\link[dplyr]{group_by} for more information of how grouping works.}

\item{verify_members}{Whether to verify the individual members of the
ensemble. Even if thresholds are supplied, only summary scores are
computed. If you wish to compute categorical scores, the separate
\link[harpPoint]{det_verify} function must be used.}

\item{show_progress}{Logical - whether to show progress bars. Defaults to
\code{TRUE}.}

\item{...}{Reserved for methods.}
}
\value{
A \code{harp_verif} object with an entry for "ens_summary_scores", and
"det_summary_scores" if \code{verify_members = TRUE}.
}
\description{
A hexbin compares the value of observations and forecasts by binning them
into hexagonal areas on a forecast vs observations hexagonal grid. By default
the score is aggregated by lead time, though any other grouping strategy can
be chosen. Hexbins for each ensemble are computed as well, but can be skipped
by setting \code{verify_members = FALSE}.
}
\details{
This is simply a wrapper around \code{\link{ens_verify()}} with \code{hexbin}
as the only selected score.
}
