% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_probabilities.R
\name{ens_probabilities}
\alias{ens_probabilities}
\title{Compute probabilities of threshold exceedence for ensemble forecasts}
\usage{
ens_probabilities(.fcst, parameter, thresholds, obs_probabilities = TRUE)
}
\arguments{
\item{.fcst}{A \code{harp_fcst} object with tables that have a column for
observations, or a single forecast table.}

\item{parameter}{The name of the column for the observed data.}

\item{thresholds}{A numeric vector of thresholds for which to compute
probabilities.}

\item{groupings}{The groups for which to compute the ensemble mean and
spread. See \link[dplyr]{group_by} for more information of how grouping
works.}
}
\value{
An object of the same format as the inputs but with data grouped for
  the \code{groupings} column(s) and columns for \code{rank} and
  \code{rank_count} that are nested together in a column with the name
  \code{rank_histogram}.
}
\description{
Compute probabilities of threshold exceedence for ensemble forecasts
}
