% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_probabilities.R
\name{ens_probabilities}
\alias{ens_probabilities}
\title{Compute probabilities of threshold exceedence for ensemble forecasts}
\usage{
ens_probabilities(.fcst, thresholds, parameter = NULL)
}
\arguments{
\item{.fcst}{A \code{harp_fcst} object with tables that have a column for
observations, or a single forecast table.}

\item{thresholds}{A numeric vector of thresholds for which to compute
probabilities.}

\item{parameter}{The name of the column for the observed data.}
}
\value{
A \code{harp_fcst} object with each data frame having columns for threshold,
fcst_prob and obs_prob instead of the columns for each member forecast.
}
\description{
Compute probabilities of threshold exceedence for ensemble forecasts
}
