% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_reliability.R
\name{ens_reliability}
\alias{ens_reliability}
\title{Reliability for an ensemble.}
\usage{
ens_reliability(
  .fcst,
  parameter,
  thresholds,
  groupings = "leadtime",
  climatology = "sample",
  rel_probs = NA,
  show_progress = TRUE,
  fcst_model = NULL,
  ...
)
}
\arguments{
\item{.fcst}{A \code{harp_fcst} object with tables that have a column for
observations, or a single forecast table.}

\item{parameter}{The name of the column for the observed data.}

\item{thresholds}{A numeric vector of thresholds for which to compute the
reliability.}

\item{groupings}{The groups for which to compute the ensemble mean and
spread. See \link[dplyr]{group_by} for more information of how grouping
works.}

\item{climatology}{The climatology to use for the Brier Skill Score. Can be
"sample" for the sample climatology (the default), a named list with
elements eps_model and member to use a member of an eps model in the
harp_fcst object for the climatology, or a data frame with columns for
threshold and climatology and also optionally leadtime.}

\item{rel_probs}{Probabilities to use for reliability diagrams. Set to NA
(the default) to select automatically.}

\item{show_progress}{Logical - whether to show a progress bar. The default is
`TRUE`}

\item{fcst_model}{The name of the forecast model to use in the `fcst_model`
column of the output. If the function is dispatched on a `harp_list`
object, the names of the `harp_list` are automatically used.}

\item{...}{Passed to \link{ens_brier}.}
}
\value{
A data frame with data grouped for the \code{groupings} column(s) and
  a nested column for reliability. The column can be unnested with
  \link[tidyr]{unnest}.
}
\description{
Reliability for an ensemble.
}
