% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_reliability.R, R/ens_score_funcs.R
\name{ens_reliability}
\alias{ens_reliability}
\title{Reliability for an ensemble.}
\usage{
ens_reliability(
  .fcst,
  parameter,
  thresholds,
  comparator = c("ge", "gt", "le", "lt", "eq", "between", "outside"),
  include_low = TRUE,
  include_high = TRUE,
  groupings = "lead_time",
  climatology = "sample",
  rel_probs = NA,
  show_progress = TRUE,
  ...
)

ens_reliability(
  .fcst,
  parameter,
  thresholds,
  comparator = c("ge", "gt", "le", "lt", "eq", "between", "outside"),
  include_low = TRUE,
  include_high = TRUE,
  groupings = "lead_time",
  climatology = "sample",
  rel_probs = NA,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{.fcst}{A \code{harp_df} or \code{harp_list} object with tables that have a column
for observations, or a single forecast table.}

\item{parameter}{The name of the column for the observations data. Can be the
column name, quoted, or unquoted. If a variable it should be embraced -
i.e. wrapped in \code{{{}}}.}

\item{thresholds}{A numeric vector of thresholds for which to compute the
threshold based scores. Set to NULL (the default) to only compute summary
scores.}

\item{groupings}{The groups for which to compute the scores. See
\link[dplyr]{group_by} for more information of how grouping works.}

\item{climatology}{The climatology to use for the Brier Skill Score. Can be
"sample" for the sample climatology (the default), a named list with
elements eps_model and member to use a member of an eps model in the
harp_fcst object for the climatology, or a data frame with columns for
threshold and climatology and also optionally lead_time.}

\item{rel_probs}{Probabilities to use for reliability diagrams. Set to NA
(the default) to select automatically.}

\item{show_progress}{Logical - whether to show progress bars. Defaults to
\code{TRUE}.}

\item{...}{Reserved for methods.}
}
\value{
A \code{harp_verif} object with an entry for "ens_threshold_scores".
}
\description{
Computes the reliability for an ensemble forecast. That is, the observed
frequency for a given forecast probability. The binning into to probability
categories can be set using \code{rel_probs}.
}
\details{
This is simply a wrapper around \code{\link{ens_verify()}} with \code{reliability}
as the only selected score.
}
