% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_roc.R
\name{ens_roc}
\alias{ens_roc}
\title{Relative Operating Characteristics for an ensemble.}
\usage{
ens_roc(
  .fcst,
  parameter,
  thresholds,
  groupings = "lead_time",
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{.fcst}{A \code{harp_fcst} object with tables that have a column for
observations, or a single forecast table.}

\item{parameter}{The name of the column for the observed data.}

\item{thresholds}{A numeric vector of thresholds for which to compute the
ROC.}

\item{groupings}{The groups for which to compute the ROC. See
\link[dplyr]{group_by} for more information of how grouping works.}

\item{show_progress}{Logical - whether to show a progress bar. The default is
`TRUE`}
}
\value{
A data frame with data grouped for the \code{groupings} column(s) and
  a nested column for the ROC with each row containing a data frame with
  columns: \code{prob} for the forecast probability bin, \code{HR} for the
  hit rate and \code{FAR} for the false alarm rate. Use \link[tidyr]{unnest}
  to unnest to the nested column.
}
\description{
Relative Operating Characteristics for an ensemble.
}
