% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_spread_and_skill.R
\name{ens_spread_and_skill}
\alias{ens_spread_and_skill}
\title{Compute the skill (RMSE) and spread of an ensemble forecast}
\usage{
ens_spread_and_skill(
  .fcst,
  parameter,
  groupings = "leadtime",
  jitter_fcst = NULL,
  ...
)
}
\arguments{
\item{.fcst}{A \code{harp_fcst} object with tables that have a column for
observations, or a single forecast table.}

\item{parameter}{The name of the column for the observed data.}

\item{groupings}{The groups for which to compute the ensemble mean and
spread. See \link[dplyr]{group_by} for more information of how grouping
works.}

\item{jitter_fcst}{A function to perturb the forecast values by. This is used
to account for observation error in the spread. For other statistics it is
likely to make little difference since it is expected that the observations
will have a mean error of zero.}

\item{...}{Not used.}
}
\value{
An object of the same format as the inputs but with data grouped for
  the \code{groupings} column(s) and columns for \code{rmse}, \code{spread}
  and \code{mean_bias}.
}
\description{
The ensemble mean and spread are computed as columns in a \code{harp_fcst}
object. Typically the scores are aggregated over lead time by other grouping
variables cam be chosen. The mean bias is also computed.
}
