% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_value.R
\name{ens_value}
\alias{ens_value}
\alias{ens_value.harp_ens_point_df}
\title{Economic value for an ensemble.}
\usage{
ens_value(
  .fcst,
  parameter,
  thresholds,
  groupings = "lead_time",
  show_progress = TRUE,
  ...
)

\method{ens_value}{harp_ens_point_df}(
  .fcst,
  parameter,
  thresholds,
  groupings = "lead_time",
  show_progress = TRUE,
  fcst_model = NULL,
  ...
)
}
\arguments{
\item{.fcst}{A \code{harp_df} or \code{harp_list} object with tables that have a column
for observations, or a single forecast table.}

\item{parameter}{The name of the column for the observations data. Can be the
column name, quoted, or unquoted. If a variable it should be embraced -
i.e. wrapped in \code{{{}}}.}

\item{thresholds}{A numeric vector of thresholds for which to compute the
threshold based scores. Set to NULL (the default) to only compute summary
scores.}

\item{groupings}{The groups for which to compute the scores. See
\link[dplyr]{group_by} for more information of how grouping works.}

\item{show_progress}{Logical - whether to show progress bars. Defaults to
\code{TRUE}.}

\item{...}{Reserved for methods.}

\item{fcst_model}{The name of the forecast model to use in the \code{fcst_model}
column of the output. If the function is dispatched on a \code{harp_list}
object, the names of the \code{harp_list} are automatically used.}
}
\value{
A data frame with data grouped for the \code{groupings} column(s) and
a nested column for the economic value with each row containing a data
frame with columns: \code{cl} for cost loss, and \code{value} for the
economic value. Use \link[tidyr]{unnest} to unnest to the nested column.
}
\description{
Economic value for an ensemble.
}
