% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_value.R
\name{ens_value}
\alias{ens_value}
\title{Economic value for an ensemble.}
\usage{
ens_value(
  .fcst,
  parameter,
  thresholds,
  groupings = "leadtime",
  show_progress = FALSE
)
}
\arguments{
\item{.fcst}{A \code{harp_fcst} object with tables that have a column for
observations, or a single forecast table.}

\item{parameter}{The name of the column for the observed data.}

\item{thresholds}{A numeric vector of thresholds for which to compute the
economic value.}

\item{groupings}{The groups for which to compute the economic value. See
\link[dplyr]{group_by} for more information of how grouping works.}
}
\value{
A data frame with data grouped for the \code{groupings} column(s) and
  a nested column for the economic value with each row containing a data
  frame with columns: \code{cl} for cost loss, and \code{value} for the
  economic value. Use \link[tidyr]{unnest} to unnest to the nested column.
}
\description{
Economic value for an ensemble.
}
