% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_verify.R
\name{ens_verify}
\alias{ens_verify}
\title{Compute all verification scores for an ensemble.}
\usage{
ens_verify(.fcst, parameter, thresholds = NULL, groupings = "leadtime")
}
\arguments{
\item{.fcst}{A \code{harp_fcst} object with tables that have a column for
observations, or a single forecast table.}

\item{parameter}{The name of the column for the observed data.}

\item{thresholds}{A numeric vector of thresholds for which to compute the
threshold based scores. Set to NULL (the default) to only compute summary
scores.}

\item{groupings}{The groups for which to compute the scores. See
\link[dplyr]{group_by} for more information of how grouping works.}
}
\value{
A list containting two data frames: \code{ens_summary_scores} and
  \code{ens_threshold_scores}.
}
\description{
Compute all verification scores for an ensemble.
}

