% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_verify.R
\name{ens_verify}
\alias{ens_verify}
\title{Compute all verification scores for an ensemble.}
\usage{
ens_verify(.fcst, parameter, thresholds = NULL, groupings = "leadtime",
  jitter_fcst = NULL, climatology = "sample", show_progress = TRUE)
}
\arguments{
\item{.fcst}{A \code{harp_fcst} object with tables that have a column for
observations, or a single forecast table.}

\item{parameter}{The name of the column for the observed data.}

\item{thresholds}{A numeric vector of thresholds for which to compute the
threshold based scores. Set to NULL (the default) to only compute summary
scores.}

\item{groupings}{The groups for which to compute the scores. See
\link[dplyr]{group_by} for more information of how grouping works.}

\item{jitter_fcst}{A function to perturb the forecast values by. This is used
to account for observation error in the rank histogram. For other
statistics it is likely to make little difference since it is expected that
the observations will have a mean error of zero.}

\item{climatology}{The climatology to use for the Brier Skill Score. Can be
"sample" for the sample climatology (the default), a named list with
elements eps_model and member to use a member of an eps model in the
harp_fcst object for the climatology, or a data frame with columns for
threshold and climatology and also optionally leadtime.}

\item{show_progress}{Logical - whether to show a progress bar. Defaults to
TRUE.}
}
\value{
A list containting two data frames: \code{ens_summary_scores} and
  \code{ens_threshold_scores}.
}
\description{
Compute all verification scores for an ensemble.
}

