% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_verify.R
\name{ens_verify}
\alias{ens_verify}
\title{Compute all verification scores for an ensemble.}
\usage{
ens_verify(
  .fcst,
  parameter,
  verify_members = TRUE,
  thresholds = NULL,
  groupings = "lead_time",
  rel_probs = NA,
  num_ref_members = NA,
  spread_drop_member = NULL,
  jitter_fcst = NULL,
  climatology = "sample",
  rank_hist = TRUE,
  crps = TRUE,
  brier = TRUE,
  roc = TRUE,
  econ_val = TRUE,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{.fcst}{A \code{harp_fcst} object with tables that have a column for
observations, or a single forecast table.}

\item{parameter}{The name of the column for the observed data.}

\item{verify_members}{Whether to verify the individual members of the
ensemble. Even if thresholds are supplied, only summary scores are
computed. If you wish to compute categorical scores, the separate
\link[harpPoint]{det_verify} function must be used.}

\item{thresholds}{A numeric vector of thresholds for which to compute the
threshold based scores. Set to NULL (the default) to only compute summary
scores.}

\item{groupings}{The groups for which to compute the scores. See
\link[dplyr]{group_by} for more information of how grouping works.}

\item{rel_probs}{Probabilities to use for reliability diagrams. Set to NA
(the default) to select automatically.}

\item{num_ref_members}{For "fair" scores, the score is scaled to be valid for
this number of ensemble members. Set to NA (the default) to not modify the
score.}

\item{spread_drop_member}{Which members to drop for the calculation of the
ensemble variance and standard deviation. For harp_fcst objects, this can
be a numeric scalar - in which case it is recycled for all forecast models;
a list or numeric vector of the same length as the harp_fcst object, or a
named list with the names corresponding to names in the harp_fcst object.}

\item{jitter_fcst}{A function to perturb the forecast values by. This is used
to account for observation error in the rank histogram. For other
statistics it is likely to make little difference since it is expected that
the observations will have a mean error of zero.}

\item{climatology}{The climatology to use for the Brier Skill Score. Can be
"sample" for the sample climatology (the default), a named list with
elements eps_model and member to use a member of an eps model in the
harp_fcst object for the climatology, or a data frame with columns for
threshold and climatology and also optionally leadtime.}

\item{show_progress}{Logical - whether to show progress bars. Defaults to
TRUE.}
}
\value{
A list containting three data frames: \code{ens_summary_scores},
  \code{ens_threshold_scores} and \code{det_summary_scores}.
}
\description{
Compute all verification scores for an ensemble.
}
