% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_members.R
\name{gather_members}
\alias{gather_members}
\title{Convert EPS forecast data from wide format data frame to long format data
frame.}
\usage{
gather_members(.fcst, member_prefix = "mbr")
}
\arguments{
\item{.fcst}{An EPS forecast data frame in wide format.}

\item{member_prefix}{Prefix for column names that contain forcasts for a
single member. The default is "mbr". Note that the column name only has to
contain \code{member_prefix}.}
}
\value{
An EPS data frame in long format.
}
\description{
Convert EPS forecast data from wide format data frame to long format data
frame.
}

