% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter_fcst.R
\name{jitter_fcst}
\alias{jitter_fcst}
\title{Jitter a forecast to account for observation errors.}
\usage{
jitter_fcst(.fcst, jitter_function, fcst_suffix = "_mbr")
}
\arguments{
\item{.fcst}{An object of class 'harp_fcst'.}

\item{jitter_function}{The function to be applied to the forecast values.}

\item{fcst_suffix}{The suffix that identifies forecast values in the
harp_fcst object. This will normally be "_mbr".}
}
\value{
An object of class harp_fcst with jittered forecast values.
}
\description{
To account for observation errors in ensemble verification the forecast
values can be perturbed by sampling from a specified error distribution.
Jittering the forecast is only likely to have an effect on the ensemble
spread and rank histograms.
}
\details{
Note that the function is applied rowwise so for sampling functions n should
always be 1.
}
\examples{
\donttest{
jitter_fcst(fcst, function(x) x + rnorm(1, ))
}
}
