% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_to_fcst.R
\name{join_to_fcst}
\alias{join_to_fcst}
\title{Combine data with point forecasts.}
\usage{
join_to_fcst(
  .fcst,
  .join,
  join_type = c("inner", "left", "right", "full", "semi", "anti"),
  by = "SID",
  force_join = FALSE,
  keep_x = TRUE,
  keep_y = FALSE,
  ...
)
}
\arguments{
\item{.fcst}{A \code{harp_fcst} object.}

\item{.join}{A data frame to join to each table in the \code{harp_fcst}
object.}

\item{join_type}{How to join the data frame. Acceptable values are: "inner",
"left", "right", "full", "semi", "anti". See \code{\link[dplyr]{join}} for
more details.}

\item{by}{Which columns to join by - if set to NULL a natural join will be
done, using all variables with common names across .fcst and .join. The
default is to join using only the SID column.}

\item{force_join}{Set to TRUE to force the join to happen even if the units
in .fcst and .join are not compatible.}

\item{keep_x, keep_y}{Where duplicate column names are found, but not used in
the join, these arguments are used to indicate whether the duplicate
columns from .fcst (\code{keep_x}), or .join (\code{keep_y}) should be
kept. The default is \code{keep_x = TRUE, keep_y = FALSE}.}

\item{...}{Other arguments for \link[dplyr]{join}.}
}
\value{
The input forecast data frame with column(s) added from \code{.join}.
}
\description{
Works in a similar way to the \link[dplyr]{join} family of functions, excpet
that the join type is specified as an argument.
}
\details{
This function would typically be used to add a column of observations to the
forecast data. If units columns are found in the data, the function will
check if the units are the same for both .fcst and .join. If they are not, an
error will be thrown and the join will not be done. This behaviour can be
overridden with the force_join argument.
}
\examples{

}
