% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_forecast.R
\name{lag_forecast}
\alias{lag_forecast}
\title{Lags a forecast}
\usage{
lag_forecast(.fcst, fcst_model, parent_cycles, direction = 1)
}
\arguments{
\item{.fcst}{A harp_fcst object.}

\item{fcst_model}{The name of the forecast model in the harp_fcst object to
be lagged. Must be quoted.}

\item{parent_cycles}{A numeric vector of forecast cycles that form the
parents for the lagging. Members between parent cycles are the child
cycles.}

\item{dierction}{The direction to do the lagging in. 1 (the default) lags
backwards in time from the parent cycles and -1 lags forwards in time.}
}
\value{
A harp_fcst object with \code{fcst_model} now containing the lagged
  forecast.
}
\description{
Lagging is done by supplying the parent forecast cycles. This function will
then work out which cycles belong to the parent and return the lagged
forecast with members from the child cycles appended to those of the parent
cycle.
}
