% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_members.R
\name{spread_members}
\alias{spread_members}
\title{Convert EPS forecast data from long format data frame to wide format data
frame.}
\usage{
spread_members(.fcst, ...)
}
\arguments{
\item{.fcst}{An EPS forecast data frame in long format.}
}
\value{
An EPS data frame in wide format.
}
\description{
Convert EPS forecast data from long format data frame to wide format data
frame.
}

